Sg1ST7qV: Correct option names and descriptions
check_jose.py:
Correct the description of the `--jose` option so that it references the fact this will be the output of the users signing implementation.

check_thumbprint.py:
Rename the `--client-signing-certificate <PATH>` to `--certificate <PATH>` and update description to reflect the fact this is used for both signing and encryption certificates.